Rename CaategoryTable to planningCategoryTable for consistency
